rename dir murtest -> masteruserrecord
