+++ pkg/controller/masteruserrecord/masteruserrecord_controller_test.go
@@ -4,7 +4,7 @@ import (
 	toolchainv1alpha1 "github.com/codeready-toolchain/api/pkg/apis/toolchain/v1alpha1"
 	"github.com/codeready-toolchain/host-operator/pkg/apis"
 	"github.com/codeready-toolchain/host-operator/test"
-	"github.com/codeready-toolchain/host-operator/test/murtest"
+	"github.com/codeready-toolchain/host-operator/test/masteruserrecord"
 	uatest "github.com/codeready-toolchain/host-operator/test/useraccount"
 	"github.com/codeready-toolchain/toolchain-common/pkg/cluster"
 	"github.com/stretchr/testify/assert"
